Move PartiallySignedTransaction to common; implement encode_partially_signed_transaction in wasm bindings
