+++ test-rpc-functions/src/rpc.rs
@@ -28,6 +28,7 @@ use common::{
             EpochIndex,
         },
         output_value::OutputValue,
+        partially_signed_transaction::PartiallySignedTransaction,
         signature::inputsig::{
             arbitrary_message,
             authorize_hashed_timelock_contract_spend::AuthorizedHashedTimelockContractSpend,
@@ -45,7 +46,6 @@ use serialization::{
     hex_encoded::HexEncoded,
     Encode as _,
 };
-use wallet_types::partially_signed_transaction::PartiallySignedTransaction;
 
 use crate::{RpcTestFunctionsError, RpcTestFunctionsHandle};
