+++ wallet/src/send_request/mod.rs
@@ -30,7 +30,7 @@ use common::primitives::{Amount, BlockHeight};
 use crypto::vrf::VRFPublicKey;
 use utils::ensure;
 use wallet_types::currency::Currency;
-use wallet_types::partially_signed_transaction::{PartiallySignedTransaction, TxAdditionalInfo};
+use wallet_types::partially_signed_transaction::{PartiallySignedTransaction, PartiallySignedTransactionWalletExt as _, TxAdditionalInfo};
 
 use crate::account::PoolData;
 use crate::destination_getters::{get_tx_output_destination, HtlcSpendingCondition};
@@ -320,7 +320,7 @@ impl SendRequest {
         let utxos = self.utxos;
         let tx = Transaction::new(self.flags, self.inputs, self.outputs)?;
 
-        let ptx = PartiallySignedTransaction::new(
+        let ptx = PartiallySignedTransaction::new_for_wallet(
             tx,
             vec![None; num_inputs],
             utxos,
